+++ socketHelper.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stddef.h>
+
 enum MessageType
 {
     MessageTerm,
@@ -12,8 +14,8 @@ enum MessageType
 
 int hostSocket(const char* port);
 int connectSocket(const char* host, const char* port);
-int recv_p(int socketFd, void* data, int numBytes);
-int send_p(int socketFd, const void* data, int numBytes);
+int recv_p(int socketFd, void* data, size_t numBytes);
+int send_p(int socketFd, const void* data, size_t numBytes);
 // must call returned char* with free()
 char* recv_str(int socketFd);
 int send_str(int socketFd, const char* string);
